+++ data/domain/src/main/java/com/escodro/domain/usecase/taskwithcategory/LoadTasksByCategory.kt
@@ -1,18 +1,12 @@
 package com.escodro.domain.usecase.taskwithcategory
 
 import com.escodro.domain.model.TaskWithCategory
-import com.escodro.domain.repository.CategoryRepository
-import com.escodro.domain.repository.TaskWithCategoryRepository
 import kotlinx.coroutines.flow.Flow
-import kotlinx.coroutines.flow.flow
 
 /**
  * Use case to get a task with category by the category id from the database.
  */
-class LoadTasksByCategory(
-    private val joinRepository: TaskWithCategoryRepository,
-    private val categoryRepository: CategoryRepository
-) {
+interface LoadTasksByCategory {
 
     /**
      * Gets a task with category by the category id if the category exists.
@@ -21,12 +15,5 @@ class LoadTasksByCategory(
      *
      * @return observable to be subscribe
      */
-    suspend operator fun invoke(categoryId: Long): Flow<List<TaskWithCategory>> {
-        val category = categoryRepository.findCategoryById(categoryId)
-        return if (category != null) {
-            joinRepository.findAllTasksWithCategoryId(category.id)
-        } else {
-            flow { throw IllegalArgumentException("Category not found with id = $categoryId") }
-        }
-    }
+    suspend operator fun invoke(categoryId: Long): Flow<List<TaskWithCategory>>
 }
